Introducing `FormattedValue` component, using to consolidate displaying decorated values.
