+++ graylog2-web-interface/src/views/components/TypeSpecificValue.tsx
@@ -16,8 +16,6 @@
  */
 import * as React from 'react';
 import { useMemo } from 'react';
-import trim from 'lodash/trim';
-import trunc from 'lodash/truncate';
 
 import FieldType from 'views/logic/fieldtypes/FieldType';
 import { getPrettifiedValue } from 'views/components/visualizations/utils/unitConverters';
@@ -27,10 +25,9 @@ import useFeature from 'hooks/useFeature';
 import { MISSING_BUCKET_NAME } from 'views/Constants';
 import formatValueWithUnitLabel from 'views/components/visualizations/utils/formatValueWithUnitLabel';
 import usePluginEntities from 'hooks/usePluginEntities';
-import stringify from 'util/stringify';
+import FormattedValue from 'views/components/FormattedValue';
 
-import EmptyValue from './EmptyValue';
-import type { ValueRendererProps, ValueRenderer } from './messagelist/decoration/ValueRenderer';
+import type { ValueRendererProps } from './messagelist/decoration/ValueRenderer';
 import DecoratorValue from './DecoratorValue';
 
 const usePluggableValueRenderer = () => {
@@ -43,17 +40,6 @@ const usePluggableValueRenderer = () => {
 };
 
 const defaultComponent = ({ value }: ValueRendererProps) => value;
-
-const _formatValue = (field: string, value: any, truncate: boolean, render: ValueRenderer, type: FieldType) => {
-  const stringified = stringify(value);
-  const Component = render;
-
-  return trim(stringified) === '' ? (
-    <EmptyValue />
-  ) : (
-    <Component field={field} value={truncate ? trunc(stringified) : stringified} type={type} />
-  );
-};
 
 type TypeSpecificValueProps = {
   field: string;
@@ -70,7 +56,7 @@ const ValueWithUnitRenderer = ({ value, unit }: { value: number; unit: FieldUnit
   return <span title={value.toString()}>{formatValueWithUnitLabel(prettified?.value, prettified.unit.abbrev)}</span>;
 };
 
-const FormattedValue = ({
+const FormattedValueWithUnits = ({
   field,
   value = undefined,
   truncate = false,
@@ -83,7 +69,7 @@ const FormattedValue = ({
     unitFeatureEnabled && unit?.isDefined && value && value !== MISSING_BUCKET_NAME && unitFeatureEnabled;
   if (shouldRenderValueWithUnit) return <ValueWithUnitRenderer value={value} unit={unit} />;
 
-  return _formatValue(field, value, truncate, render, type);
+  return <FormattedValue field={field} value={value} truncate={truncate} render={render} type={type} />;
 };
 
 const TypeSpecificValue = ({
@@ -108,7 +94,9 @@ const TypeSpecificValue = ({
     return pluggableValueRenderer[type.type](value, field, render);
   }
 
-  return <FormattedValue field={field} value={value} truncate={truncate} unit={unit} render={render} type={type} />;
+  return (
+    <FormattedValueWithUnits field={field} value={value} truncate={truncate} unit={unit} render={render} type={type} />
+  );
 };
 
 export default TypeSpecificValue;
